requirements.txt
requests>=2.28.0
numpy>=1.21.0
numpy-stl>=2.16.0
tqdm>=4.62.0
